+++ apps/api/src/utils/validators/number_validator.py
@@ -1,0 +1,6 @@
+"""Number validation utilities."""
+
+
+def validate_positive_number(value: float) -> bool:
+    """Check if a number is positive."""
+    return value > 0
